third_party/WebKit/LayoutTests/http/tests/security/xssAuditor/window-open-block-mode-expected.txt

CONSOLE ERROR: line 4: The XSS Auditor blocked access to 'http://127.0.0.1:8000/security/xssAuditor/resources/echo-intertag.pl?q=%3Cscript%3Ealert(/xss/)%3C/script%3E&enable-full-block=1' because the source code of a script was found within the request. The server sent an 'X-XSS-Protection' header requesting this behavior.
Test that under x-xss-protection mode=block that the block page URL is added on top of the existing page instead of replacing it. Test passes if the second window has a two-item back-forward list.



============== Back Forward List ==============
curr->  http://127.0.0.1:8000/security/xssAuditor/window-open-block-mode.html
===============================================

============== Back Forward List ==============
        http://127.0.0.1:8000/security/xssAuditor/resources/echo-intertag.pl?q=%3Cscript%3Ealert(/xss/)%3C/script%3E&enable-full-block=1
curr->  data:,
===============================================
